PyPathomics-main_v2/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
cachetools==6.1.0
category-encoders==2.6.4
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.0.9
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.2
fonttools==4.59.1
formulaic==1.2.0
gprofiler-official==1.0.0
graphviz==0.21
greenlet==3.1.1
gseapy==1.1.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
interface-meta==1.3.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langserve==0.3.1
langsmith==0.4.12
large-image==1.32.11
large-image-source-openslide==1.32.11
large-image-source-pil==1.32.11
large-image-source-tiff==1.32.11
lazy-property==0.0.1
lazy_loader==0.4
lifelines==0.30.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mrmr-selection==0.2.8
msgpack==1.1.1
narwhals==2.1.2
nbformat==5.10.4
networkx==3.2.1
numpy==1.24.4
openai==1.100.2
opencv-python==4.8.1.78
openpyxl==3.1.5
openslide-bin==4.0.0.8
openslide-python==1.4.2
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
palettable==3.3.3
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.49.1
plotly==6.3.0
polars==1.32.3
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyee==12.0.0
Pygments==2.19.2
pykwalify==1.8.0
pylibtiff==0.6.1
PyMuPDF==1.26.4
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
ridgeplot==0.3.2
rpds-py==0.27.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
simpleitk==2.5.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
statsmodels==0.14.5
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.8.30
tifftools==1.6.1
tiktoken==0.11.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
UpSetPlot==0.9.0
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.3
xgboost==2.1.4
xxhash==3.5.0
zipp==3.23.0
zstandard==0.24.0
